+++ NEWS.md
@@ -4,6 +4,11 @@
 
 * `tableby()`, `paired()`: Add warning for when by-variable contains empty string. (#121)
 
+* `tableby()`, `modelsum()`: properly propogated "term.name" to the `as.data.frame()` method. (#127, #128)
+
+* `tableby()`, `modelsum()`: Allow for LaTeX formatting. NOTE: this changes the formatting behavior when specifying
+  `text="html"`! (#123)
+
 # arsenal v1.3.0
 
 **This is a mostly backwards-compatible update.**